+++ src/merkleblock.h
@@ -52,7 +52,7 @@
 class CPartialMerkleTree {
 protected:
     /** the total number of transactions in the block */
-    unsigned int nTransactions;
+    uint32_t nTransactions;
 
     /** node-is-parent-of-matched-txid bits */
     std::vector<bool> vBits;
@@ -67,7 +67,7 @@ protected:
      * Helper function to efficiently calculate the number of nodes at given
      * height in the merkle tree.
      */
-    unsigned int CalcTreeWidth(int height) const {
+    size_t CalcTreeWidth(int height) const {
         return (nTransactions + (1 << height) - 1) >> height;
     }
 
@@ -75,14 +75,13 @@ protected:
      * Calculate the hash of a node in the merkle tree (at leaf level: the
      * txid's themselves)
      */
-    uint256 CalcHash(int height, unsigned int pos,
-                     const std::vector<uint256> &vTxid);
+    uint256 CalcHash(int height, size_t pos, const std::vector<uint256> &vTxid);
 
     /**
      * Recursive function that traverses tree nodes, storing the data as bits
      * and hashes.
      */
-    void TraverseAndBuild(int height, unsigned int pos,
+    void TraverseAndBuild(int height, size_t pos,
                           const std::vector<uint256> &vTxid,
                           const std::vector<bool> &vMatch);
 
@@ -91,10 +90,9 @@ protected:
      * hashes produced by TraverseAndBuild. It returns the hash of the
      * respective node and its respective index.
      */
-    uint256 TraverseAndExtract(int height, unsigned int pos,
-                               unsigned int &nBitsUsed, unsigned int &nHashUsed,
-                               std::vector<uint256> &vMatch,
-                               std::vector<unsigned int> &vnIndex);
+    uint256 TraverseAndExtract(int height, size_t pos, size_t &nBitsUsed,
+                               size_t &nHashUsed, std::vector<uint256> &vMatch,
+                               std::vector<size_t> &vnIndex);
 
 public:
     /** serialization implementation */
@@ -109,13 +107,13 @@ public:
             READWRITE(vBytes);
             CPartialMerkleTree &us = *(const_cast<CPartialMerkleTree *>(this));
             us.vBits.resize(vBytes.size() * 8);
-            for (unsigned int p = 0; p < us.vBits.size(); p++) {
+            for (size_t p = 0; p < us.vBits.size(); p++) {
                 us.vBits[p] = (vBytes[p / 8] & (1 << (p % 8))) != 0;
             }
             us.fBad = false;
         } else {
             vBytes.resize((vBits.size() + 7) / 8);
-            for (unsigned int p = 0; p < vBits.size(); p++) {
+            for (size_t p = 0; p < vBits.size(); p++) {
                 vBytes[p / 8] |= vBits[p] << (p % 8);
             }
             READWRITE(vBytes);
@@ -137,7 +135,7 @@ public:
      * root, or 0 in case of failure.
      */
     uint256 ExtractMatches(std::vector<uint256> &vMatch,
-                           std::vector<unsigned int> &vnIndex);
+                           std::vector<size_t> &vnIndex);
 };
 
 /**
@@ -159,7 +157,7 @@ public:
     CPartialMerkleTree txn;
 
     /** Public only for unit testing and relay testing (not relayed) */
-    std::vector<std::pair<unsigned int, uint256>> vMatchedTxn;
+    std::vector<std::pair<size_t, uint256>> vMatchedTxn;
 
     /**
      * Create a Merkle proof according to a bloom filter. Note
