+++ src/test/pmt_tests.cpp
@@ -87,7 +87,7 @@ BOOST_AUTO_TEST_CASE(pmt_test1) {
 
             // extract merkle root and matched txids from copy
             std::vector<uint256> vMatchTxid2;
-            std::vector<unsigned int> vIndex;
+            std::vector<size_t> vIndex;
             uint256 merkleRoot2 = pmt2.ExtractMatches(vMatchTxid2, vIndex);
 
             // check that it has the same merkle root as the original, and a
@@ -121,7 +121,7 @@ BOOST_AUTO_TEST_CASE(pmt_malleability) {
                                 false, false, false, true,  true,  false};
 
     CPartialMerkleTree tree(vTxid, vMatch);
-    std::vector<unsigned int> vIndex;
+    std::vector<size_t> vIndex;
     BOOST_CHECK(tree.ExtractMatches(vTxid, vIndex).IsNull());
 }
